+++ src/constants/filterConfig.js
@@ -2,6 +2,10 @@ import { buildDateRangeQuery } from '../utils';
 
 export const filterConfig = [
   {
+    name: 'createdDate',
+    parse: buildDateRangeQuery('createdDate'),
+  },
+  {
     name: 'updatedDate',
     parse: buildDateRangeQuery('updatedDate'),
   },